# Remove getval from vspace(getval(...)) pattern (depends on #260)