--- a/autograd/util.py
+++ b/autograd/util.py
@@ -1,14 +1,11 @@
 from __future__ import absolute_import
 from __future__ import print_function
 from copy import copy
-from operator import itemgetter
-from future.utils import iteritems
-from builtins import map, range, zip
+from builtins import range
 
 import autograd.numpy as np
 from autograd.convenience_wrappers import grad
 from autograd.core import vspace, vspace_flatten, getval
-from autograd.container_types import make_tuple, make_list, make_dict
 
 EPS, RTOL, ATOL = 1e-4, 1e-4, 1e-6