This is a code simplification.

Includes an edit to `assert_vspace_match`, which is part of our 'inner loop', however, this doesn't appear to produce any performance regressions.

Benchmarking this branch against master, the only timings to change significantly are those of `flatten`, as expected (this is because of the changes in https://github.com/HIPS/autograd/pull/260, which this branch is based on). There is a slight regression for `vspace_flatten`, which was discussed in https://github.com/HIPS/autograd/pull/260 and is acceptable IMO since this method is only used by the tests.

```
All benchmarks:

    before     after       ratio
  [723e6777] [823c048d]
     1.14μs     1.05μs      0.92  bench_core.time_exp_call
     3.88μs     4.00μs      1.03  bench_core.time_exp_primitive_call_boxed
     3.02μs     3.21μs      1.06  bench_core.time_exp_primitive_call_unboxed
   624.39ms   592.63ms      0.95  bench_core.time_fan_out_fan_in_backward_pass
   229.18ms   230.75ms      1.01  bench_core.time_fan_out_fan_in_forward_pass
   847.61ms   831.57ms      0.98  bench_core.time_fan_out_fan_in_grad
   838.28μs   827.38μs      0.99  bench_core.time_long_backward_pass
   413.39μs   428.94μs      1.04  bench_core.time_long_forward_pass
     1.31ms     1.29ms      0.98  bench_core.time_long_grad
     3.77μs     3.75μs      0.99  bench_core.time_new_node_array
     3.63μs     3.47μs      0.96  bench_core.time_new_node_float
    39.90μs    37.92μs      0.95  bench_core.time_short_backward_pass
    18.08μs    18.29μs      1.01  bench_core.time_short_forward_pass
    90.82μs    87.40μs      0.96  bench_core.time_short_grad
     2.59μs     2.56μs      0.99  bench_core.time_vspace_array
     2.56μs     2.58μs      1.01  bench_core.time_vspace_float
      1.97y      1.96y      1.00  bench_rnn.RNNSuite.peakmem_manual_rnn_grad
      2.09y      2.10y      1.00  bench_rnn.RNNSuite.peakmem_rnn_grad
   609.16ms   617.79ms      1.01  bench_rnn.RNNSuite.time_manual_rnn_grad
   664.51ms   667.79ms      1.00  bench_rnn.RNNSuite.time_rnn_grad
-  290.68μs   211.05μs      0.73  bench_util.time_flatten
-    1.51ms   645.68μs      0.43  bench_util.time_grad_flatten
+   78.13μs   106.64μs      1.36  bench_util.time_vspace_flatten
```