@j-towns 2017-07-19
OK sorry for being a git noob, it seems this pr includes the commits from https://github.com/HIPS/autograd/pull/260 as well as the ones which I intended. If you first merge https://github.com/HIPS/autograd/pull/260 maybe github will be smart enough to only include [the _extra_ commits in this pr](https://github.com/j-towns/autograd/compare/vspace-nodes...j-towns:remove-vspace-getval)?