@dougalm 2017-07-25
Looks great to me. Just rebased onto master and merged.